refactor: adiciona tipos de retorno
