+++ unbook/backend/package.json
@@ -26,11 +26,14 @@
   },
   "homepage": "https://github.com/mdsreq-fga-unb/2022.2-UnBook#readme",
   "dependencies": {
+    "-": "^0.0.1",
+    "@faker-js/faker": "^7.6.0",
     "@types/bcrypt": "^5.0.0",
     "@types/express": "^4.17.15",
     "@types/supertest": "^2.0.12",
     "@types/typescript": "^2.0.0",
     "bcrypt": "^5.1.0",
+    "D": "^1.0.0",
     "express": "^4.18.2",
     "i": "^0.3.7",
     "jsonwebtoken": "^9.0.0",
