+++ unbook/backend/tests/presentation/signup/email-validator-adapter.spec.ts
@@ -1,5 +1,5 @@
 import validator from "validator";
-import { EmailValidatorAdapter } from "../../src/infra/validators/EmailValidatorAdapter";
+import { EmailValidatorAdapter } from "../../../src/infra/validators/EmailValidatorAdapter";
 
 jest.mock("validator", () => {
   return {
